refactor(test): ExceptionAssertions를 test-support로 통합
각 모듈에 산재한 coffeeshout.fixture.ExceptionAssertions 복사본(8개)을
test-support의 coffeeshout.support.ExceptionAssertions 단일 소스로 통합.
:user, :game, :admin, :websocket, :zzolbot에 testImplementation(:test-support) 추가.
