+++ backend/admin/build.gradle.kts
@@ -1,5 +1,9 @@
 // :admin — 운영자 영역 (dashboard + patchnote + report)
 
+plugins {
+    `java-test-fixtures`
+}
+
 tasks.bootJar { enabled = false }
 tasks.jar { enabled = true }
 
@@ -32,4 +36,7 @@ dependencies {
     annotationProcessor("com.querydsl:querydsl-apt:$queryDslVersion:jakarta")
     annotationProcessor("jakarta.annotation:jakarta.annotation-api")
     annotationProcessor("jakarta.persistence:jakarta.persistence-api")
+
+    testFixturesImplementation(project(":game-api"))
+    testImplementation(project(":test-support"))
 }
