+++ backend/app/src/test/java/coffeeshout/blockstacking/application/BlockStackingServiceTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.blockstacking.application;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.never;
