+++ backend/app/src/test/java/coffeeshout/friend/application/service/FriendshipServiceTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.friend.application.service;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.mockito.ArgumentMatchers.any;
