+++ backend/app/src/test/java/coffeeshout/user/application/service/AuthTokenServiceTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.user.application.service;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.fixture.UserFixture;
