+++ backend/game/build.gradle.kts
@@ -1,5 +1,9 @@
 // :game — 6게임 + orchestration (이전 minigame)
 
+plugins {
+    `java-test-fixtures`
+}
+
 tasks.bootJar { enabled = false }
 tasks.jar { enabled = true }
 
@@ -21,4 +25,7 @@ dependencies {
     annotationProcessor("com.querydsl:querydsl-apt:$queryDslVersion:jakarta")
     annotationProcessor("jakarta.annotation:jakarta.annotation-api")
     annotationProcessor("jakarta.persistence:jakarta.persistence-api")
+
+    testFixturesImplementation(project(":game-api"))
+    testImplementation(project(":test-support"))
 }
