+++ backend/game/src/test/java/coffeeshout/cardgame/domain/CardGameTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.cardgame.domain;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
