+++ backend/game/src/test/java/coffeeshout/laddergame/domain/BottomRanksTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.laddergame.domain;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Collection;
