+++ backend/room/src/test/java/coffeeshout/room/domain/RoomTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.room.domain;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
