+++ backend/room/src/test/java/coffeeshout/room/domain/roulette/ProbabilityCalculatorTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.room.domain.roulette;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.room.domain.RoomErrorCode;
