+++ backend/room/src/test/java/coffeeshout/room/domain/service/PlayerNameGeneratorTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.room.domain.service;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.global.nickname.WordPicker;
