+++ backend/room/src/test/java/coffeeshout/room/infra/PlayerNameAuditPromptTemplateTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.room.infra;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
