+++ backend/room/src/test/java/coffeeshout/room/infra/auth/JjwtRoomSessionTokenIssuerTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.Test;
 
 import java.time.Duration;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
